Preparation in order to allow thread pool / parallel tasks in vcmi
- Moved ThreadPool from RMG to main library.
- Removed ancient version of ThreadPool
- Simplified ThreadPool usage
