+++ AI/BattleAI/BattleAI.cpp
@@ -16,7 +16,6 @@
 #include "EnemyInfo.h"
 #include "tbb/parallel_for.h"
 #include "../../lib/CStopWatch.h"
-#include "../../lib/CThreadHelper.h"
 #include "../../lib/mapObjects/CGTownInstance.h"
 #include "../../lib/spells/CSpellHandler.h"
 #include "../../lib/spells/ISpellMechanics.h"
