+++ AI/Nullkiller/AIGateway.cpp
@@ -17,6 +17,7 @@
 #include "../../lib/GameSettings.h"
 #include "../../lib/gameState/CGameState.h"
 #include "../../lib/NetPacks.h"
+#include "../../lib/ThreadUtilities.h"
 #include "../../lib/serializer/CTypeList.h"
 #include "../../lib/serializer/BinarySerializer.h"
 #include "../../lib/serializer/BinaryDeserializer.h"
