+++ AI/VCAI/VCAI.h
@@ -14,8 +14,6 @@
 #include "../../lib/AI_Base.h"
 #include "../../CCallback.h"
 
-#include "../../lib/CThreadHelper.h"
-
 #include "../../lib/GameConstants.h"
 #include "../../lib/VCMI_Lib.h"
 #include "../../lib/CBuildingHandler.h"
