+++ lib/ThreadUtilities.h
@@ -1,0 +1,16 @@
+/*
+ * CThreadHelper.h, part of VCMI engine
+ *
+ * Authors: listed in file AUTHORS in main folder
+ *
+ * License: GNU General Public License v2.0 or later
+ * Full text of license available in license.txt file, in main folder
+ *
+ */
+#pragma once
+
+VCMI_LIB_NAMESPACE_BEGIN
+
+void DLL_LINKAGE setThreadName(const std::string &name);
+
+VCMI_LIB_NAMESPACE_END
