+++ lib/rmg/RmgMap.h
@@ -11,7 +11,7 @@
 #pragma once
 #include "../int3.h"
 #include "../GameConstants.h"
-#include "threadpool/MapProxy.h"
+#include "MapProxy.h"
 
 VCMI_LIB_NAMESPACE_BEGIN
